"Updated type signature of partialGeneral function in TT.hs"
Model used: gpt-4-0613

me> That's not quite right, so I've added a comment with some guidance, can you take a look and see if you can update the type signature based on that comment, please?
jr>  I 've  updated  the  type  signature  of  the  ` partial General `  function  in  the  ` src / TT .hs `  file .  The  function  is  currently  undefined ,  but  its  type  signature  is  now  ` forall  a  b  c .  For all  a  Un con strained 1  =>  Template  (( a  . +  b )  . +  c )  ->  Rec  ( a  . +  b )  ->  Template  c `.  This  should  allow  for  handling  a  case  where  we  have  a  nested  row -type  and  we  pass  only  part  of  the  outer  type  containing  only  part  of  the  inner  type .